@@ -4,11 +4,11 @@ description: A Flutter plugin for getting information about and controlling the
   and streaming image buffers to dart.
 repository: https://github.com/flutter/plugins/tree/master/packages/camera/camera
 issue_tracker: https://github.com/flutter/flutter/issues?q=is%3Aissue+is%3Aopen+label%3A%22p%3A+camera%22
-version: 0.9.0
+version: 0.9.3
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.0.0"
+  sdk: ">=2.14.0 <3.0.0"
+  flutter: ">=2.5.0"
 
 flutter:
   plugin:
@@ -20,7 +20,7 @@ flutter:
         pluginClass: CameraPlugin
 
 dependencies:
-  camera_platform_interface: ^2.0.0
+  camera_platform_interface: ^2.1.0
   flutter:
     sdk: flutter
   pedantic: ^1.10.0
