@@ -399,6 +399,13 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
                             onSetExposureModeButtonPressed(ExposureMode.locked)
                         : null,
                   ),
+                  TextButton(
+                    child: Text('RESET OFFSET'),
+                    style: styleLocked,
+                    onPressed: controller != null
+                        ? () => controller!.setExposureOffset(0.0)
+                        : null,
+                  ),
                 ],
               ),
               Center(
@@ -530,7 +537,16 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
                   cameraController.value.isRecordingVideo
               ? onStopButtonPressed
               : null,
-        )
+        ),
+        IconButton(
+          icon: const Icon(Icons.pause_presentation),
+          color:
+              cameraController != null && cameraController.value.isPreviewPaused
+                  ? Colors.red
+                  : Colors.blue,
+          onPressed:
+              cameraController == null ? null : onPausePreviewButtonPressed,
+        ),
       ],
     );
   }
@@ -594,7 +610,9 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
   }
 
   void onNewCameraSelected(CameraDescription cameraDescription) async {
-    final previousCameraController = controller;
+    if (controller != null) {
+      await controller!.dispose();
+    }
 
     final CameraController cameraController = CameraController(
       cameraDescription,
@@ -604,10 +622,6 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
     );
 
     controller = cameraController;
-
-    if (mounted) {
-      setState(() {});
-    }
 
     // If the controller is updated then update the UI.
     cameraController.addListener(() {
@@ -641,8 +655,6 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
     if (mounted) {
       setState(() {});
     }
-
-    await previousCameraController?.dispose();
   }
 
   void onTakePictureButtonPressed() {
@@ -745,6 +757,23 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
         _startVideoPlayer();
       }
     });
+  }
+
+  Future<void> onPausePreviewButtonPressed() async {
+    final CameraController? cameraController = controller;
+
+    if (cameraController == null || !cameraController.value.isInitialized) {
+      showInSnackBar('Error: select a camera first.');
+      return;
+    }
+
+    if (cameraController.value.isPreviewPaused) {
+      await cameraController.resumePreview();
+    } else {
+      await cameraController.pausePreview();
+    }
+
+    if (mounted) setState(() {});
   }
 
   void onPauseButtonPressed() {
