@@ -5,6 +5,7 @@
 import 'dart:html';
 import 'dart:ui';
 
+import 'package:async/async.dart';
 import 'package:camera_platform_interface/camera_platform_interface.dart';
 import 'package:camera_web/src/camera.dart';
 import 'package:camera_web/src/camera_service.dart';
@@ -84,11 +85,17 @@ void main() {
           'creates a video element '
           'with correct properties', (tester) async {
         const audioConstraints = AudioConstraints(enabled: true);
+        final videoConstraints = VideoConstraints(
+          facingMode: FacingModeConstraint(
+            CameraType.user,
+          ),
+        );
 
         final camera = Camera(
           textureId: textureId,
           options: CameraOptions(
             audio: audioConstraints,
+            video: videoConstraints,
           ),
           cameraService: cameraService,
         );
@@ -97,7 +104,7 @@ void main() {
 
         expect(camera.videoElement, isNotNull);
         expect(camera.videoElement.autoplay, isFalse);
-        expect(camera.videoElement.muted, !audioConstraints.enabled);
+        expect(camera.videoElement.muted, isTrue);
         expect(camera.videoElement.srcObject, mediaStream);
         expect(camera.videoElement.attributes.keys, contains('playsinline'));
 
@@ -107,6 +114,27 @@ void main() {
         expect(camera.videoElement.style.width, equals('100%'));
         expect(camera.videoElement.style.height, equals('100%'));
         expect(camera.videoElement.style.objectFit, equals('cover'));
+      });
+
+      testWidgets(
+          'flips the video element horizontally '
+          'for a back camera', (tester) async {
+        final videoConstraints = VideoConstraints(
+          facingMode: FacingModeConstraint(
+            CameraType.environment,
+          ),
+        );
+
+        final camera = Camera(
+          textureId: textureId,
+          options: CameraOptions(
+            video: videoConstraints,
+          ),
+          cameraService: cameraService,
+        );
+
+        await camera.initialize();
+
         expect(camera.videoElement.style.transform, equals('scaleX(-1)'));
       });
 
@@ -375,7 +403,7 @@ void main() {
         await camera.initialize();
 
         expect(
-          await camera.getVideoSize(),
+          camera.getVideoSize(),
           equals(videoSize),
         );
       });
@@ -395,7 +423,7 @@ void main() {
         await camera.initialize();
 
         expect(
-          await camera.getVideoSize(),
+          camera.getVideoSize(),
           equals(Size.zero),
         );
       });
@@ -818,6 +846,87 @@ void main() {
       });
     });
 
+    group('getLensDirection', () {
+      testWidgets(
+          'returns a lens direction '
+          'based on the first video track settings', (tester) async {
+        final videoElement = MockVideoElement();
+
+        final camera = Camera(
+          textureId: textureId,
+          cameraService: cameraService,
+        )..videoElement = videoElement;
+
+        final firstVideoTrack = MockMediaStreamTrack();
+
+        when(() => videoElement.srcObject).thenReturn(
+          FakeMediaStream([
+            firstVideoTrack,
+            MockMediaStreamTrack(),
+          ]),
+        );
+
+        when(firstVideoTrack.getSettings)
+            .thenReturn({'facingMode': 'environment'});
+
+        when(() => cameraService.mapFacingModeToLensDirection('environment'))
+            .thenReturn(CameraLensDirection.external);
+
+        expect(
+          camera.getLensDirection(),
+          equals(CameraLensDirection.external),
+        );
+      });
+
+      testWidgets(
+          'returns null '
+          'if the first video track is missing the facing mode',
+          (tester) async {
+        final videoElement = MockVideoElement();
+
+        final camera = Camera(
+          textureId: textureId,
+          cameraService: cameraService,
+        )..videoElement = videoElement;
+
+        final firstVideoTrack = MockMediaStreamTrack();
+
+        when(() => videoElement.srcObject).thenReturn(
+          FakeMediaStream([
+            firstVideoTrack,
+            MockMediaStreamTrack(),
+          ]),
+        );
+
+        when(firstVideoTrack.getSettings).thenReturn({});
+
+        expect(
+          camera.getLensDirection(),
+          isNull,
+        );
+      });
+
+      testWidgets(
+          'returns null '
+          'if the camera is missing video tracks', (tester) async {
+        // Create a video stream with no video tracks.
+        final videoElement = VideoElement();
+        mediaStream = videoElement.captureStream();
+
+        final camera = Camera(
+          textureId: textureId,
+          cameraService: cameraService,
+        );
+
+        await camera.initialize();
+
+        expect(
+          camera.getLensDirection(),
+          isNull,
+        );
+      });
+    });
+
     group('getViewType', () {
       testWidgets('returns a correct view type', (tester) async {
         final camera = Camera(
@@ -843,9 +952,80 @@ void main() {
 
         await camera.initialize();
 
-        camera.dispose();
+        await camera.dispose();
 
         expect(camera.videoElement.srcObject, isNull);
+      });
+    });
+
+    group('events', () {
+      group('onEnded', () {
+        testWidgets(
+            'emits the default video track '
+            'when it emits an ended event', (tester) async {
+          final camera = Camera(
+            textureId: textureId,
+            cameraService: cameraService,
+          );
+
+          final streamQueue = StreamQueue(camera.onEnded);
+
+          await camera.initialize();
+
+          final videoTracks = camera.stream!.getVideoTracks();
+          final defaultVideoTrack = videoTracks.first;
+
+          defaultVideoTrack.dispatchEvent(Event('ended'));
+
+          expect(
+            await streamQueue.next,
+            equals(defaultVideoTrack),
+          );
+
+          await streamQueue.cancel();
+        });
+
+        testWidgets(
+            'emits the default video track '
+            'when the camera is stopped', (tester) async {
+          final camera = Camera(
+            textureId: textureId,
+            cameraService: cameraService,
+          );
+
+          final streamQueue = StreamQueue(camera.onEnded);
+
+          await camera.initialize();
+
+          final videoTracks = camera.stream!.getVideoTracks();
+          final defaultVideoTrack = videoTracks.first;
+
+          camera.stop();
+
+          expect(
+            await streamQueue.next,
+            equals(defaultVideoTrack),
+          );
+
+          await streamQueue.cancel();
+        });
+
+        testWidgets(
+            'no longer emits the default video track '
+            'when the camera is disposed', (tester) async {
+          final camera = Camera(
+            textureId: textureId,
+            cameraService: cameraService,
+          );
+
+          await camera.initialize();
+          await camera.dispose();
+
+          expect(
+            camera.onEndedStreamController.isClosed,
+            isTrue,
+          );
+        });
       });
     });
   });
