@@ -13,6 +13,7 @@ import 'package:platform/platform.dart';
 import 'core.dart';
 import 'plugin_command.dart';
 import 'process_runner.dart';
+import 'repository_package.dart';
 
 /// Possible outcomes of a command run for a package.
 enum RunState {
@@ -84,12 +85,24 @@ abstract class PackageLoopingCommand extends PluginCommand {
   int _otherWarningCount = 0;
 
   /// The package currently being run by [runForPackage].
-  PackageEnumerationEntry? _currentPackage;
+  PackageEnumerationEntry? _currentPackageEntry;
 
   /// Called during [run] before any calls to [runForPackage]. This provides an
   /// opportunity to fail early if the command can't be run (e.g., because the
   /// arguments are invalid), and to set up any run-level state.
   Future<void> initializeRun() async {}
+
+  /// Returns the packages to process. By default, this returns the packages
+  /// defined by the standard tooling flags and the [inculdeSubpackages] option,
+  /// but can be overridden for custom package enumeration.
+  ///
+  /// Note: Consistent behavior across commands whenever possibel is a goal for
+  /// this tool, so this should be overridden only in rare cases.
+  Stream<PackageEnumerationEntry> getPackagesToProcess() async* {
+    yield* includeSubpackages
+        ? getTargetPackagesAndSubpackages(filterExcluded: false)
+        : getTargetPackages(filterExcluded: false);
+  }
 
   /// Runs the command for [package], returning a list of errors.
   ///
@@ -97,7 +110,7 @@ abstract class PackageLoopingCommand extends PluginCommand {
   /// be included in the final error summary (e.g., a command that only has a
   /// single failure mode), or strings that should be listed for that package
   /// in the final summary. An empty list indicates success.
-  Future<PackageResult> runForPackage(Directory package);
+  Future<PackageResult> runForPackage(RepositoryPackage package);
 
   /// Called during [run] after all calls to [runForPackage]. This provides an
   /// opportunity to do any cleanup of run-level state.
@@ -137,6 +150,9 @@ abstract class PackageLoopingCommand extends PluginCommand {
   /// context.
   String get failureListFooter => 'See above for full details.';
 
+  /// The summary string used for a successful run in the final overview output.
+  String get successSummaryMessage => 'ran';
+
   /// If true, all printing (including the summary) will be redirected to a
   /// buffer, and provided in a call to [handleCapturedOutput] at the end of
   /// the run.
@@ -155,29 +171,11 @@ abstract class PackageLoopingCommand extends PluginCommand {
   /// things that might be useful to someone debugging an unexpected result.
   void logWarning(String warningMessage) {
     print(Colorize(warningMessage)..yellow());
-    if (_currentPackage != null) {
-      _packagesWithWarnings.add(_currentPackage!);
+    if (_currentPackageEntry != null) {
+      _packagesWithWarnings.add(_currentPackageEntry!);
     } else {
       ++_otherWarningCount;
     }
-  }
-
-  /// Returns the identifying name to use for [package].
-  ///
-  /// Implementations should not expect a specific format for this string, since
-  /// it uses heuristics to try to be precise without being overly verbose. If
-  /// an exact format (e.g., published name, or basename) is required, that
-  /// should be used instead.
-  String getPackageDescription(Directory package) {
-    String packageName = getRelativePosixPath(package, from: packagesDir);
-    final List<String> components = p.posix.split(packageName);
-    // For the common federated plugin pattern of `foo/foo_subpackage`, drop
-    // the first part since it's not useful.
-    if (components.length >= 2 &&
-        components[1].startsWith('${components[0]}_')) {
-      packageName = p.posix.joinAll(components.sublist(1));
-    }
-    return packageName;
   }
 
   /// Returns the relative path from [from] to [entity] in Posix style.
@@ -219,36 +217,42 @@ abstract class PackageLoopingCommand extends PluginCommand {
   Future<bool> _runInternal() async {
     _packagesWithWarnings.clear();
     _otherWarningCount = 0;
-    _currentPackage = null;
+    _currentPackageEntry = null;
 
     await initializeRun();
 
-    final List<PackageEnumerationEntry> packages = includeSubpackages
-        ? await getTargetPackagesAndSubpackages(filterExcluded: false).toList()
-        : await getTargetPackages(filterExcluded: false).toList();
+    final List<PackageEnumerationEntry> targetPackages =
+        await getPackagesToProcess().toList();
 
     final Map<PackageEnumerationEntry, PackageResult> results =
         <PackageEnumerationEntry, PackageResult>{};
-    for (final PackageEnumerationEntry package in packages) {
-      _currentPackage = package;
-      _printPackageHeading(package);
+    for (final PackageEnumerationEntry entry in targetPackages) {
+      _currentPackageEntry = entry;
+      _printPackageHeading(entry);
 
       // Command implementations should never see excluded packages; they are
       // included at this level only for logging.
-      if (package.excluded) {
-        results[package] = PackageResult.exclude();
+      if (entry.excluded) {
+        results[entry] = PackageResult.exclude();
         continue;
       }
 
-      final PackageResult result = await runForPackage(package.directory);
+      PackageResult result;
+      try {
+        result = await runForPackage(entry.package);
+      } catch (e, stack) {
+        printError(e.toString());
+        printError(stack.toString());
+        result = PackageResult.fail(<String>['Unhandled exception']);
+      }
       if (result.state == RunState.skipped) {
         final String message =
             '${indentation}SKIPPING: ${result.details.first}';
         captureOutput ? print(message) : print(Colorize(message)..darkGray());
       }
-      results[package] = result;
+      results[entry] = result;
     }
-    _currentPackage = null;
+    _currentPackageEntry = null;
 
     completeRun();
 
@@ -256,13 +260,13 @@ abstract class PackageLoopingCommand extends PluginCommand {
     // If there were any errors reported, summarize them and exit.
     if (results.values
         .any((PackageResult result) => result.state == RunState.failed)) {
-      _printFailureSummary(packages, results);
+      _printFailureSummary(targetPackages, results);
       return false;
     }
 
     // Otherwise, print a summary of what ran for ease of auditing that all the
     // expected tests ran.
-    _printRunSummary(packages, results);
+    _printRunSummary(targetPackages, results);
 
     print('\n');
     _printSuccess('No issues found!');
@@ -283,9 +287,9 @@ abstract class PackageLoopingCommand extends PluginCommand {
   /// Something is always printed to make it easier to distinguish between
   /// a command running for a package and producing no output, and a command
   /// not having been run for a package.
-  void _printPackageHeading(PackageEnumerationEntry package) {
-    final String packageDisplayName = getPackageDescription(package.directory);
-    String heading = package.excluded
+  void _printPackageHeading(PackageEnumerationEntry entry) {
+    final String packageDisplayName = entry.package.displayName;
+    String heading = entry.excluded
         ? 'Not running for $packageDisplayName; excluded'
         : 'Running for $packageDisplayName';
     if (hasLongOutput) {
@@ -295,16 +299,15 @@ abstract class PackageLoopingCommand extends PluginCommand {
 || $heading
 ============================================================
 ''';
-    } else if (!package.excluded) {
+    } else if (!entry.excluded) {
       heading = '$heading...';
     }
     if (captureOutput) {
       print(heading);
     } else {
       final Colorize colorizeHeading = Colorize(heading);
-      print(package.excluded
-          ? colorizeHeading.darkGray()
-          : colorizeHeading.cyan());
+      print(
+          entry.excluded ? colorizeHeading.darkGray() : colorizeHeading.cyan());
     }
   }
 
@@ -349,21 +352,22 @@ abstract class PackageLoopingCommand extends PluginCommand {
 
   /// Prints a one-line-per-package overview of the run results for each
   /// package.
-  void _printPerPackageRunOverview(List<PackageEnumerationEntry> packages,
+  void _printPerPackageRunOverview(
+      List<PackageEnumerationEntry> packageEnumeration,
       {required Set<PackageEnumerationEntry> skipped}) {
     print('Run overview:');
-    for (final PackageEnumerationEntry package in packages) {
-      final bool hadWarning = _packagesWithWarnings.contains(package);
+    for (final PackageEnumerationEntry entry in packageEnumeration) {
+      final bool hadWarning = _packagesWithWarnings.contains(entry);
       Styles style;
       String summary;
-      if (package.excluded) {
+      if (entry.excluded) {
         summary = 'excluded';
         style = Styles.DARK_GRAY;
-      } else if (skipped.contains(package)) {
+      } else if (skipped.contains(entry)) {
         summary = 'skipped';
         style = hadWarning ? Styles.LIGHT_YELLOW : Styles.DARK_GRAY;
       } else {
-        summary = 'ran';
+        summary = successSummaryMessage;
         style = hadWarning ? Styles.YELLOW : Styles.GREEN;
       }
       if (hadWarning) {
@@ -373,18 +377,18 @@ abstract class PackageLoopingCommand extends PluginCommand {
       if (!captureOutput) {
         summary = (Colorize(summary)..apply(style)).toString();
       }
-      print('  ${getPackageDescription(package.directory)} - $summary');
+      print('  ${entry.package.displayName} - $summary');
     }
     print('');
   }
 
   /// Prints a summary of all of the failures from [results].
-  void _printFailureSummary(List<PackageEnumerationEntry> packages,
+  void _printFailureSummary(List<PackageEnumerationEntry> packageEnumeration,
       Map<PackageEnumerationEntry, PackageResult> results) {
     const String indentation = '  ';
     _printError(failureListHeader);
-    for (final PackageEnumerationEntry package in packages) {
-      final PackageResult result = results[package]!;
+    for (final PackageEnumerationEntry entry in packageEnumeration) {
+      final PackageResult result = results[entry]!;
       if (result.state == RunState.failed) {
         final String errorIndentation = indentation * 2;
         String errorDetails = '';
@@ -392,8 +396,7 @@ abstract class PackageLoopingCommand extends PluginCommand {
           errorDetails =
               ':\n$errorIndentation${result.details.join('\n$errorIndentation')}';
         }
-        _printError(
-            '$indentation${getPackageDescription(package.directory)}$errorDetails');
+        _printError('$indentation${entry.package.displayName}$errorDetails');
       }
     }
     _printError(failureListFooter);
