@@ -11,6 +11,7 @@ import 'package:pubspec_parse/pubspec_parse.dart';
 
 import 'common/core.dart';
 import 'common/plugin_command.dart';
+import 'common/repository_package.dart';
 
 const String _outputDirectoryFlag = 'output-dir';
 
@@ -170,10 +171,11 @@ class CreateAllPluginsAppCommand extends PluginCommand {
     final Map<String, PathDependency> pathDependencies =
         <String, PathDependency>{};
 
-    await for (final PackageEnumerationEntry package in getTargetPackages()) {
+    await for (final PackageEnumerationEntry entry in getTargetPackages()) {
+      final RepositoryPackage package = entry.package;
       final Directory pluginDirectory = package.directory;
       final String pluginName = pluginDirectory.basename;
-      final File pubspecFile = pluginDirectory.childFile('pubspec.yaml');
+      final File pubspecFile = package.pubspecFile;
       final Pubspec pubspec = Pubspec.parse(pubspecFile.readAsStringSync());
 
       if (pubspec.publishTo != 'none') {
