@@ -12,6 +12,7 @@ import 'common/core.dart';
 import 'common/package_looping_command.dart';
 import 'common/plugin_utils.dart';
 import 'common/process_runner.dart';
+import 'common/repository_package.dart';
 
 const int _exitNoPlatformFlags = 2;
 const int _exitNoAvailableDevice = 3;
@@ -35,7 +36,10 @@ class DriveExamplesCommand extends PackageLoopingCommand {
     argParser.addFlag(kPlatformWeb,
         help: 'Runs the web implementation of the examples');
     argParser.addFlag(kPlatformWindows,
-        help: 'Runs the Windows implementation of the examples');
+        help: 'Runs the Windows (Win32) implementation of the examples');
+    argParser.addFlag(kPlatformWinUwp,
+        help:
+            'Runs the UWP implementation of the examples [currently a no-op]');
     argParser.addOption(
       kEnableExperiment,
       defaultsTo: '',
@@ -66,6 +70,7 @@ class DriveExamplesCommand extends PackageLoopingCommand {
       kPlatformMacos,
       kPlatformWeb,
       kPlatformWindows,
+      kPlatformWinUwp,
     ];
     final int platformCount = platformSwitches
         .where((String platform) => getBoolArg(platform))
@@ -78,6 +83,10 @@ class DriveExamplesCommand extends PackageLoopingCommand {
           'Exactly one of ${platformSwitches.map((String platform) => '--$platform').join(', ')} '
           'must be specified.');
       throw ToolExit(_exitNoPlatformFlags);
+    }
+
+    if (getBoolArg(kPlatformWinUwp)) {
+      logWarning('Driving UWP applications is not yet supported');
     }
 
     String? androidDevice;
@@ -115,13 +124,17 @@ class DriveExamplesCommand extends PackageLoopingCommand {
         ],
       if (getBoolArg(kPlatformWindows))
         kPlatformWindows: <String>['-d', 'windows'],
+      // TODO(stuartmorgan): Check these flags once drive supports UWP:
+      // https://github.com/flutter/flutter/issues/82821
+      if (getBoolArg(kPlatformWinUwp))
+        kPlatformWinUwp: <String>['-d', 'winuwp'],
     };
   }
 
   @override
-  Future<PackageResult> runForPackage(Directory package) async {
-    if (package.basename.endsWith('_platform_interface') &&
-        !package.childDirectory('example').existsSync()) {
+  Future<PackageResult> runForPackage(RepositoryPackage package) async {
+    if (package.directory.basename.endsWith('_platform_interface') &&
+        !package.getSingleExampleDeprecated().directory.existsSync()) {
       // Platform interface packages generally aren't intended to have
       // examples, and don't need integration tests, so skip rather than fail.
       return PackageResult.skip(
@@ -131,7 +144,17 @@ class DriveExamplesCommand extends PackageLoopingCommand {
     final List<String> deviceFlags = <String>[];
     for (final MapEntry<String, List<String>> entry
         in _targetDeviceFlags.entries) {
-      if (pluginSupportsPlatform(entry.key, package)) {
+      final String platform = entry.key;
+      String? variant;
+      if (platform == kPlatformWindows) {
+        variant = platformVariantWin32;
+      } else if (platform == kPlatformWinUwp) {
+        variant = platformVariantWinUwp;
+        // TODO(stuartmorgan): Remove this once drive supports UWP.
+        // https://github.com/flutter/flutter/issues/82821
+        return PackageResult.skip('Drive does not yet support UWP');
+      }
+      if (pluginSupportsPlatform(platform, package, variant: variant)) {
         deviceFlags.addAll(entry.value);
       } else {
         print('Skipping unsupported platform ${entry.key}...');
@@ -140,16 +163,16 @@ class DriveExamplesCommand extends PackageLoopingCommand {
     // If there is no supported target platform, skip the plugin.
     if (deviceFlags.isEmpty) {
       return PackageResult.skip(
-          '${getPackageDescription(package)} does not support any requested platform.');
+          '${package.displayName} does not support any requested platform.');
     }
 
     int examplesFound = 0;
     bool testsRan = false;
     final List<String> errors = <String>[];
-    for (final Directory example in getExamplesForPlugin(package)) {
+    for (final RepositoryPackage example in package.getExamples()) {
       ++examplesFound;
       final String exampleName =
-          getRelativePosixPath(example, from: packagesDir);
+          getRelativePosixPath(example.directory, from: packagesDir);
 
       final List<File> drivers = await _getDrivers(example);
       if (drivers.isEmpty) {
@@ -173,7 +196,7 @@ class DriveExamplesCommand extends PackageLoopingCommand {
 
         if (testTargets.isEmpty) {
           final String driverRelativePath =
-              getRelativePosixPath(driver, from: package);
+              getRelativePosixPath(driver, from: package.directory);
           printError(
               'Found $driverRelativePath, but no integration_test/*_test.dart files.');
           errors.add('No test files for $driverRelativePath');
@@ -185,7 +208,8 @@ class DriveExamplesCommand extends PackageLoopingCommand {
             example, driver, testTargets,
             deviceFlags: deviceFlags);
         for (final File failingTarget in failingTargets) {
-          errors.add(getRelativePosixPath(failingTarget, from: package));
+          errors.add(
+              getRelativePosixPath(failingTarget, from: package.directory));
         }
       }
     }
@@ -229,10 +253,10 @@ class DriveExamplesCommand extends PackageLoopingCommand {
     return deviceIds;
   }
 
-  Future<List<File>> _getDrivers(Directory example) async {
+  Future<List<File>> _getDrivers(RepositoryPackage example) async {
     final List<File> drivers = <File>[];
 
-    final Directory driverDir = example.childDirectory('test_driver');
+    final Directory driverDir = example.directory.childDirectory('test_driver');
     if (driverDir.existsSync()) {
       await for (final FileSystemEntity driver in driverDir.list()) {
         if (driver is File && driver.basename.endsWith('_test.dart')) {
@@ -253,10 +277,10 @@ class DriveExamplesCommand extends PackageLoopingCommand {
     return testFile.existsSync() ? testFile : null;
   }
 
-  Future<List<File>> _getIntegrationTests(Directory example) async {
+  Future<List<File>> _getIntegrationTests(RepositoryPackage example) async {
     final List<File> tests = <File>[];
     final Directory integrationTestDir =
-        example.childDirectory('integration_test');
+        example.directory.childDirectory('integration_test');
 
     if (integrationTestDir.existsSync()) {
       await for (final FileSystemEntity file in integrationTestDir.list()) {
@@ -278,7 +302,7 @@ class DriveExamplesCommand extends PackageLoopingCommand {
   ///   - `['-d', 'web-server', '--web-port=<port>', '--browser-name=<browser>]`
   ///     for web
   Future<List<File>> _driveTests(
-    Directory example,
+    RepositoryPackage example,
     File driver,
     List<File> targets, {
     required List<String> deviceFlags,
@@ -296,11 +320,11 @@ class DriveExamplesCommand extends PackageLoopingCommand {
             if (enableExperiment.isNotEmpty)
               '--enable-experiment=$enableExperiment',
             '--driver',
-            getRelativePosixPath(driver, from: example),
+            getRelativePosixPath(driver, from: example.directory),
             '--target',
-            getRelativePosixPath(target, from: example),
+            getRelativePosixPath(target, from: example.directory),
           ],
-          workingDir: example);
+          workingDir: example.directory);
       if (exitCode != 0) {
         failures.add(target);
       }
