@@ -7,6 +7,7 @@ import 'dart:io';
 import 'package:args/command_runner.dart';
 import 'package:file/file.dart';
 import 'package:file/memory.dart';
+import 'package:flutter_plugin_tools/src/common/core.dart';
 import 'package:flutter_plugin_tools/src/common/plugin_command.dart';
 import 'package:flutter_plugin_tools/src/common/process_runner.dart';
 import 'package:git/git.dart';
@@ -28,8 +29,6 @@ void main() {
   late MockPlatform mockPlatform;
   late Directory packagesDir;
   late Directory thirdPartyPackagesDir;
-  late List<List<String>?> gitDirCommands;
-  late String gitDiffResponse;
 
   setUp(() {
     fileSystem = MemoryFileSystem();
@@ -39,18 +38,15 @@ void main() {
         .childDirectory('third_party')
         .childDirectory('packages');
 
-    gitDirCommands = <List<String>?>[];
-    gitDiffResponse = '';
     final MockGitDir gitDir = MockGitDir();
     when(gitDir.runCommand(any, throwOnError: anyNamed('throwOnError')))
         .thenAnswer((Invocation invocation) {
-      gitDirCommands.add(invocation.positionalArguments[0] as List<String>?);
-      final MockProcessResult mockProcessResult = MockProcessResult();
-      if (invocation.positionalArguments[0][0] == 'diff') {
-        when<String?>(mockProcessResult.stdout as String?)
-            .thenReturn(gitDiffResponse);
-      }
-      return Future<ProcessResult>.value(mockProcessResult);
+      final List<String> arguments =
+          invocation.positionalArguments[0]! as List<String>;
+      // Attach the first argument to the command to make targeting the mock
+      // results easier.
+      final String gitCommand = arguments.removeAt(0);
+      return processRunner.run('git-$gitCommand', arguments);
     });
     processRunner = RecordingProcessRunner();
     command = SamplePluginCommand(
@@ -184,6 +180,68 @@ void main() {
       expect(command.plugins, unorderedEquals(<String>[]));
     });
 
+    group('conflicting package selection', () {
+      test('does not allow --packages with --run-on-changed-packages',
+          () async {
+        Error? commandError;
+        final List<String> output = await runCapturingPrint(runner, <String>[
+          'sample',
+          '--run-on-changed-packages',
+          '--packages=plugin1',
+        ], errorHandler: (Error e) {
+          commandError = e;
+        });
+
+        expect(commandError, isA<ToolExit>());
+        expect(
+            output,
+            containsAllInOrder(<Matcher>[
+              contains('Only one of --packages, --run-on-changed-packages, or '
+                  '--packages-for-branch can be provided.')
+            ]));
+      });
+
+      test('does not allow --packages with --packages-for-branch', () async {
+        Error? commandError;
+        final List<String> output = await runCapturingPrint(runner, <String>[
+          'sample',
+          '--packages-for-branch',
+          '--packages=plugin1',
+        ], errorHandler: (Error e) {
+          commandError = e;
+        });
+
+        expect(commandError, isA<ToolExit>());
+        expect(
+            output,
+            containsAllInOrder(<Matcher>[
+              contains('Only one of --packages, --run-on-changed-packages, or '
+                  '--packages-for-branch can be provided.')
+            ]));
+      });
+
+      test(
+          'does not allow --run-on-changed-packages with --packages-for-branch',
+          () async {
+        Error? commandError;
+        final List<String> output = await runCapturingPrint(runner, <String>[
+          'sample',
+          '--packages-for-branch',
+          '--packages=plugin1',
+        ], errorHandler: (Error e) {
+          commandError = e;
+        });
+
+        expect(commandError, isA<ToolExit>());
+        expect(
+            output,
+            containsAllInOrder(<Matcher>[
+              contains('Only one of --packages, --run-on-changed-packages, or '
+                  '--packages-for-branch can be provided.')
+            ]));
+      });
+    });
+
     group('test run-on-changed-packages', () {
       test('all plugins should be tested if there are no changes.', () async {
         final Directory plugin1 = createFakePlugin('plugin1', packagesDir);
@@ -201,7 +259,9 @@ void main() {
       test(
           'all plugins should be tested if there are no plugin related changes.',
           () async {
-        gitDiffResponse = 'AUTHORS';
+        processRunner.mockProcessesForExecutable['git-diff'] = <Process>[
+          MockProcess(stdout: 'AUTHORS'),
+        ];
         final Directory plugin1 = createFakePlugin('plugin1', packagesDir);
         final Directory plugin2 = createFakePlugin('plugin2', packagesDir);
         await runCapturingPrint(runner, <String>[
@@ -215,10 +275,12 @@ void main() {
       });
 
       test('all plugins should be tested if .cirrus.yml changes.', () async {
-        gitDiffResponse = '''
+        processRunner.mockProcessesForExecutable['git-diff'] = <Process>[
+          MockProcess(stdout: '''
 .cirrus.yml
 packages/plugin1/CHANGELOG
-''';
+'''),
+        ];
         final Directory plugin1 = createFakePlugin('plugin1', packagesDir);
         final Directory plugin2 = createFakePlugin('plugin2', packagesDir);
         await runCapturingPrint(runner, <String>[
@@ -232,10 +294,12 @@ packages/plugin1/CHANGELOG
       });
 
       test('all plugins should be tested if .ci.yaml changes', () async {
-        gitDiffResponse = '''
+        processRunner.mockProcessesForExecutable['git-diff'] = <Process>[
+          MockProcess(stdout: '''
 .ci.yaml
 packages/plugin1/CHANGELOG
-''';
+'''),
+        ];
         final Directory plugin1 = createFakePlugin('plugin1', packagesDir);
         final Directory plugin2 = createFakePlugin('plugin2', packagesDir);
         await runCapturingPrint(runner, <String>[
@@ -250,10 +314,12 @@ packages/plugin1/CHANGELOG
 
       test('all plugins should be tested if anything in .ci/ changes',
           () async {
-        gitDiffResponse = '''
+        processRunner.mockProcessesForExecutable['git-diff'] = <Process>[
+          MockProcess(stdout: '''
 .ci/Dockerfile
 packages/plugin1/CHANGELOG
-''';
+'''),
+        ];
         final Directory plugin1 = createFakePlugin('plugin1', packagesDir);
         final Directory plugin2 = createFakePlugin('plugin2', packagesDir);
         await runCapturingPrint(runner, <String>[
@@ -268,10 +334,12 @@ packages/plugin1/CHANGELOG
 
       test('all plugins should be tested if anything in script changes.',
           () async {
-        gitDiffResponse = '''
+        processRunner.mockProcessesForExecutable['git-diff'] = <Process>[
+          MockProcess(stdout: '''
 script/tool_runner.sh
 packages/plugin1/CHANGELOG
-''';
+'''),
+        ];
         final Directory plugin1 = createFakePlugin('plugin1', packagesDir);
         final Directory plugin2 = createFakePlugin('plugin2', packagesDir);
         await runCapturingPrint(runner, <String>[
@@ -286,10 +354,12 @@ packages/plugin1/CHANGELOG
 
       test('all plugins should be tested if the root analysis options change.',
           () async {
-        gitDiffResponse = '''
+        processRunner.mockProcessesForExecutable['git-diff'] = <Process>[
+          MockProcess(stdout: '''
 analysis_options.yaml
 packages/plugin1/CHANGELOG
-''';
+'''),
+        ];
         final Directory plugin1 = createFakePlugin('plugin1', packagesDir);
         final Directory plugin2 = createFakePlugin('plugin2', packagesDir);
         await runCapturingPrint(runner, <String>[
@@ -304,10 +374,12 @@ packages/plugin1/CHANGELOG
 
       test('all plugins should be tested if formatting options change.',
           () async {
-        gitDiffResponse = '''
+        processRunner.mockProcessesForExecutable['git-diff'] = <Process>[
+          MockProcess(stdout: '''
 .clang-format
 packages/plugin1/CHANGELOG
-''';
+'''),
+        ];
         final Directory plugin1 = createFakePlugin('plugin1', packagesDir);
         final Directory plugin2 = createFakePlugin('plugin2', packagesDir);
         await runCapturingPrint(runner, <String>[
@@ -321,24 +393,35 @@ packages/plugin1/CHANGELOG
       });
 
       test('Only changed plugin should be tested.', () async {
-        gitDiffResponse = 'packages/plugin1/plugin1.dart';
+        processRunner.mockProcessesForExecutable['git-diff'] = <Process>[
+          MockProcess(stdout: 'packages/plugin1/plugin1.dart'),
+        ];
         final Directory plugin1 = createFakePlugin('plugin1', packagesDir);
         createFakePlugin('plugin2', packagesDir);
-        await runCapturingPrint(runner, <String>[
+        final List<String> output = await runCapturingPrint(runner, <String>[
           'sample',
           '--base-sha=master',
           '--run-on-changed-packages'
         ]);
+
+        expect(
+            output,
+            containsAllInOrder(<Matcher>[
+              contains(
+                  'Running for all packages that have changed relative to "master"'),
+            ]));
 
         expect(command.plugins, unorderedEquals(<String>[plugin1.path]));
       });
 
       test('multiple files in one plugin should also test the plugin',
           () async {
-        gitDiffResponse = '''
+        processRunner.mockProcessesForExecutable['git-diff'] = <Process>[
+          MockProcess(stdout: '''
 packages/plugin1/plugin1.dart
 packages/plugin1/ios/plugin1.m
-''';
+'''),
+        ];
         final Directory plugin1 = createFakePlugin('plugin1', packagesDir);
         createFakePlugin('plugin2', packagesDir);
         await runCapturingPrint(runner, <String>[
@@ -352,10 +435,12 @@ packages/plugin1/ios/plugin1.m
 
       test('multiple plugins changed should test all the changed plugins',
           () async {
-        gitDiffResponse = '''
+        processRunner.mockProcessesForExecutable['git-diff'] = <Process>[
+          MockProcess(stdout: '''
 packages/plugin1/plugin1.dart
 packages/plugin2/ios/plugin2.m
-''';
+'''),
+        ];
         final Directory plugin1 = createFakePlugin('plugin1', packagesDir);
         final Directory plugin2 = createFakePlugin('plugin2', packagesDir);
         createFakePlugin('plugin3', packagesDir);
@@ -372,11 +457,13 @@ packages/plugin2/ios/plugin2.m
       test(
           'multiple plugins inside the same plugin group changed should output the plugin group name',
           () async {
-        gitDiffResponse = '''
+        processRunner.mockProcessesForExecutable['git-diff'] = <Process>[
+          MockProcess(stdout: '''
 packages/plugin1/plugin1/plugin1.dart
 packages/plugin1/plugin1_platform_interface/plugin1_platform_interface.dart
 packages/plugin1/plugin1_web/plugin1_web.dart
-''';
+'''),
+        ];
         final Directory plugin1 =
             createFakePlugin('plugin1', packagesDir.childDirectory('plugin1'));
         createFakePlugin('plugin2', packagesDir);
@@ -393,9 +480,11 @@ packages/plugin1/plugin1_web/plugin1_web.dart
       test(
           'changing one plugin in a federated group should include all plugins in the group',
           () async {
-        gitDiffResponse = '''
+        processRunner.mockProcessesForExecutable['git-diff'] = <Process>[
+          MockProcess(stdout: '''
 packages/plugin1/plugin1/plugin1.dart
-''';
+'''),
+        ];
         final Directory plugin1 =
             createFakePlugin('plugin1', packagesDir.childDirectory('plugin1'));
         final Directory plugin2 = createFakePlugin('plugin1_platform_interface',
@@ -414,35 +503,14 @@ packages/plugin1/plugin1/plugin1.dart
                 <String>[plugin1.path, plugin2.path, plugin3.path]));
       });
 
-      test(
-          '--packages flag overrides the behavior of --run-on-changed-packages',
-          () async {
-        gitDiffResponse = '''
-packages/plugin1/plugin1.dart
-packages/plugin2/ios/plugin2.m
-packages/plugin3/plugin3.dart
-''';
-        final Directory plugin1 =
-            createFakePlugin('plugin1', packagesDir.childDirectory('plugin1'));
-        final Directory plugin2 = createFakePlugin('plugin2', packagesDir);
-        createFakePlugin('plugin3', packagesDir);
-        await runCapturingPrint(runner, <String>[
-          'sample',
-          '--packages=plugin1,plugin2',
-          '--base-sha=master',
-          '--run-on-changed-packages'
-        ]);
-
-        expect(command.plugins,
-            unorderedEquals(<String>[plugin1.path, plugin2.path]));
-      });
-
       test('--exclude flag works with --run-on-changed-packages', () async {
-        gitDiffResponse = '''
+        processRunner.mockProcessesForExecutable['git-diff'] = <Process>[
+          MockProcess(stdout: '''
 packages/plugin1/plugin1.dart
 packages/plugin2/ios/plugin2.m
 packages/plugin3/plugin3.dart
-''';
+'''),
+        ];
         final Directory plugin1 =
             createFakePlugin('plugin1', packagesDir.childDirectory('plugin1'));
         createFakePlugin('plugin2', packagesDir);
@@ -456,6 +524,74 @@ packages/plugin3/plugin3.dart
 
         expect(command.plugins, unorderedEquals(<String>[plugin1.path]));
       });
+    });
+  });
+
+  group('--packages-for-branch', () {
+    test('only tests changed packages on a branch', () async {
+      processRunner.mockProcessesForExecutable['git-diff'] = <Process>[
+        MockProcess(stdout: 'packages/plugin1/plugin1.dart'),
+      ];
+      processRunner.mockProcessesForExecutable['git-rev-parse'] = <Process>[
+        MockProcess(stdout: 'a-branch'),
+      ];
+      final Directory plugin1 = createFakePlugin('plugin1', packagesDir);
+      createFakePlugin('plugin2', packagesDir);
+
+      final List<String> output = await runCapturingPrint(
+          runner, <String>['sample', '--packages-for-branch']);
+
+      expect(command.plugins, unorderedEquals(<String>[plugin1.path]));
+      expect(
+          output,
+          containsAllInOrder(<Matcher>[
+            contains('--packages-for-branch: running on changed packages'),
+          ]));
+    });
+
+    test('tests all packages on master', () async {
+      processRunner.mockProcessesForExecutable['git-diff'] = <Process>[
+        MockProcess(stdout: 'packages/plugin1/plugin1.dart'),
+      ];
+      processRunner.mockProcessesForExecutable['git-rev-parse'] = <Process>[
+        MockProcess(stdout: 'master'),
+      ];
+      final Directory plugin1 = createFakePlugin('plugin1', packagesDir);
+      final Directory plugin2 = createFakePlugin('plugin2', packagesDir);
+
+      final List<String> output = await runCapturingPrint(
+          runner, <String>['sample', '--packages-for-branch']);
+
+      expect(command.plugins,
+          unorderedEquals(<String>[plugin1.path, plugin2.path]));
+      expect(
+          output,
+          containsAllInOrder(<Matcher>[
+            contains('--packages-for-branch: running on all packages'),
+          ]));
+    });
+
+    test('throws if getting the branch fails', () async {
+      processRunner.mockProcessesForExecutable['git-diff'] = <Process>[
+        MockProcess(stdout: 'packages/plugin1/plugin1.dart'),
+      ];
+      processRunner.mockProcessesForExecutable['git-rev-parse'] = <Process>[
+        MockProcess(exitCode: 1),
+      ];
+
+      Error? commandError;
+      final List<String> output = await runCapturingPrint(
+          runner, <String>['sample', '--packages-for-branch'],
+          errorHandler: (Error e) {
+        commandError = e;
+      });
+
+      expect(commandError, isA<ToolExit>());
+      expect(
+          output,
+          containsAllInOrder(<Matcher>[
+            contains('Unabled to determine branch'),
+          ]));
     });
   });
 
@@ -498,7 +634,7 @@ packages/plugin3/plugin3.dart
         expect(
             localCommand.plugins,
             unorderedEquals(expectedShards[i]
-                .map((Directory package) => package.path)
+                .map((Directory packageDir) => packageDir.path)
                 .toList()));
       }
     });
@@ -541,7 +677,7 @@ packages/plugin3/plugin3.dart
         expect(
             localCommand.plugins,
             unorderedEquals(expectedShards[i]
-                .map((Directory package) => package.path)
+                .map((Directory packageDir) => packageDir.path)
                 .toList()));
       }
     });
@@ -594,7 +730,7 @@ packages/plugin3/plugin3.dart
         expect(
             localCommand.plugins,
             unorderedEquals(expectedShards[i]
-                .map((Directory package) => package.path)
+                .map((Directory packageDir) => packageDir.path)
                 .toList()));
       }
     });
@@ -620,10 +756,8 @@ class SamplePluginCommand extends PluginCommand {
 
   @override
   Future<void> run() async {
-    await for (final PackageEnumerationEntry package in getTargetPackages()) {
-      plugins.add(package.directory.path);
+    await for (final PackageEnumerationEntry entry in getTargetPackages()) {
+      plugins.add(entry.package.path);
     }
   }
 }
-
-class MockProcessResult extends Mock implements ProcessResult {}
