@@ -6,6 +6,7 @@ import 'package:file/file.dart';
 import 'package:file/memory.dart';
 import 'package:flutter_plugin_tools/src/common/core.dart';
 import 'package:flutter_plugin_tools/src/common/plugin_utils.dart';
+import 'package:flutter_plugin_tools/src/common/repository_package.dart';
 import 'package:test/test.dart';
 
 import '../util.dart';
@@ -21,7 +22,8 @@ void main() {
 
   group('pluginSupportsPlatform', () {
     test('no platforms', () async {
-      final Directory plugin = createFakePlugin('plugin', packagesDir);
+      final RepositoryPackage plugin =
+          RepositoryPackage(createFakePlugin('plugin', packagesDir));
 
       expect(pluginSupportsPlatform(kPlatformAndroid, plugin), isFalse);
       expect(pluginSupportsPlatform(kPlatformIos, plugin), isFalse);
@@ -32,15 +34,16 @@ void main() {
     });
 
     test('all platforms', () async {
-      final Directory plugin = createFakePlugin('plugin', packagesDir,
-          platformSupport: <String, PlatformSupport>{
-            kPlatformAndroid: PlatformSupport.inline,
-            kPlatformIos: PlatformSupport.inline,
-            kPlatformLinux: PlatformSupport.inline,
-            kPlatformMacos: PlatformSupport.inline,
-            kPlatformWeb: PlatformSupport.inline,
-            kPlatformWindows: PlatformSupport.inline,
-          });
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
+          'plugin', packagesDir,
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.inline),
+            kPlatformIos: const PlatformDetails(PlatformSupport.inline),
+            kPlatformLinux: const PlatformDetails(PlatformSupport.inline),
+            kPlatformMacos: const PlatformDetails(PlatformSupport.inline),
+            kPlatformWeb: const PlatformDetails(PlatformSupport.inline),
+            kPlatformWindows: const PlatformDetails(PlatformSupport.inline),
+          }));
 
       expect(pluginSupportsPlatform(kPlatformAndroid, plugin), isTrue);
       expect(pluginSupportsPlatform(kPlatformIos, plugin), isTrue);
@@ -51,15 +54,13 @@ void main() {
     });
 
     test('some platforms', () async {
-      final Directory plugin = createFakePlugin(
-        'plugin',
-        packagesDir,
-        platformSupport: <String, PlatformSupport>{
-          kPlatformAndroid: PlatformSupport.inline,
-          kPlatformLinux: PlatformSupport.inline,
-          kPlatformWeb: PlatformSupport.inline,
-        },
-      );
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
+          'plugin', packagesDir,
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.inline),
+            kPlatformLinux: const PlatformDetails(PlatformSupport.inline),
+            kPlatformWeb: const PlatformDetails(PlatformSupport.inline),
+          }));
 
       expect(pluginSupportsPlatform(kPlatformAndroid, plugin), isTrue);
       expect(pluginSupportsPlatform(kPlatformIos, plugin), isFalse);
@@ -70,18 +71,16 @@ void main() {
     });
 
     test('inline plugins are only detected as inline', () async {
-      final Directory plugin = createFakePlugin(
-        'plugin',
-        packagesDir,
-        platformSupport: <String, PlatformSupport>{
-          kPlatformAndroid: PlatformSupport.inline,
-          kPlatformIos: PlatformSupport.inline,
-          kPlatformLinux: PlatformSupport.inline,
-          kPlatformMacos: PlatformSupport.inline,
-          kPlatformWeb: PlatformSupport.inline,
-          kPlatformWindows: PlatformSupport.inline,
-        },
-      );
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
+          'plugin', packagesDir,
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.inline),
+            kPlatformIos: const PlatformDetails(PlatformSupport.inline),
+            kPlatformLinux: const PlatformDetails(PlatformSupport.inline),
+            kPlatformMacos: const PlatformDetails(PlatformSupport.inline),
+            kPlatformWeb: const PlatformDetails(PlatformSupport.inline),
+            kPlatformWindows: const PlatformDetails(PlatformSupport.inline),
+          }));
 
       expect(
           pluginSupportsPlatform(kPlatformAndroid, plugin,
@@ -134,19 +133,16 @@ void main() {
     });
 
     test('federated plugins are only detected as federated', () async {
-      const String pluginName = 'plugin';
-      final Directory plugin = createFakePlugin(
-        pluginName,
-        packagesDir,
-        platformSupport: <String, PlatformSupport>{
-          kPlatformAndroid: PlatformSupport.federated,
-          kPlatformIos: PlatformSupport.federated,
-          kPlatformLinux: PlatformSupport.federated,
-          kPlatformMacos: PlatformSupport.federated,
-          kPlatformWeb: PlatformSupport.federated,
-          kPlatformWindows: PlatformSupport.federated,
-        },
-      );
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
+          'plugin', packagesDir,
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.federated),
+            kPlatformIos: const PlatformDetails(PlatformSupport.federated),
+            kPlatformLinux: const PlatformDetails(PlatformSupport.federated),
+            kPlatformMacos: const PlatformDetails(PlatformSupport.federated),
+            kPlatformWeb: const PlatformDetails(PlatformSupport.federated),
+            kPlatformWindows: const PlatformDetails(PlatformSupport.federated),
+          }));
 
       expect(
           pluginSupportsPlatform(kPlatformAndroid, plugin,
@@ -196,6 +192,153 @@ void main() {
           pluginSupportsPlatform(kPlatformWindows, plugin,
               requiredMode: PlatformSupport.inline),
           isFalse);
+    });
+
+    test('windows without variants is only win32', () async {
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
+        'plugin',
+        packagesDir,
+        platformSupport: <String, PlatformDetails>{
+          kPlatformWindows: const PlatformDetails(PlatformSupport.inline),
+        },
+      ));
+
+      expect(
+          pluginSupportsPlatform(kPlatformWindows, plugin,
+              variant: platformVariantWin32),
+          isTrue);
+      expect(
+          pluginSupportsPlatform(kPlatformWindows, plugin,
+              variant: platformVariantWinUwp),
+          isFalse);
+    });
+
+    test('windows with both variants matches win32 and winuwp', () async {
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
+          'plugin', packagesDir,
+          platformSupport: <String, PlatformDetails>{
+            kPlatformWindows: const PlatformDetails(
+              PlatformSupport.federated,
+              variants: <String>[platformVariantWin32, platformVariantWinUwp],
+            ),
+          }));
+
+      expect(
+          pluginSupportsPlatform(kPlatformWindows, plugin,
+              variant: platformVariantWin32),
+          isTrue);
+      expect(
+          pluginSupportsPlatform(kPlatformWindows, plugin,
+              variant: platformVariantWinUwp),
+          isTrue);
+    });
+
+    test('win32 plugin is only win32', () async {
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
+          'plugin', packagesDir,
+          platformSupport: <String, PlatformDetails>{
+            kPlatformWindows: const PlatformDetails(
+              PlatformSupport.federated,
+              variants: <String>[platformVariantWin32],
+            ),
+          }));
+
+      expect(
+          pluginSupportsPlatform(kPlatformWindows, plugin,
+              variant: platformVariantWin32),
+          isTrue);
+      expect(
+          pluginSupportsPlatform(kPlatformWindows, plugin,
+              variant: platformVariantWinUwp),
+          isFalse);
+    });
+
+    test('winup plugin is only winuwp', () async {
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
+        'plugin',
+        packagesDir,
+        platformSupport: <String, PlatformDetails>{
+          kPlatformWindows: const PlatformDetails(PlatformSupport.federated,
+              variants: <String>[platformVariantWinUwp]),
+        },
+      ));
+
+      expect(
+          pluginSupportsPlatform(kPlatformWindows, plugin,
+              variant: platformVariantWin32),
+          isFalse);
+      expect(
+          pluginSupportsPlatform(kPlatformWindows, plugin,
+              variant: platformVariantWinUwp),
+          isTrue);
+    });
+  });
+
+  group('pluginHasNativeCodeForPlatform', () {
+    test('returns false for web', () async {
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
+        'plugin',
+        packagesDir,
+        platformSupport: <String, PlatformDetails>{
+          kPlatformWeb: const PlatformDetails(PlatformSupport.inline),
+        },
+      ));
+
+      expect(pluginHasNativeCodeForPlatform(kPlatformWeb, plugin), isFalse);
+    });
+
+    test('returns false for a native-only plugin', () async {
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
+        'plugin',
+        packagesDir,
+        platformSupport: <String, PlatformDetails>{
+          kPlatformLinux: const PlatformDetails(PlatformSupport.inline),
+          kPlatformMacos: const PlatformDetails(PlatformSupport.inline),
+          kPlatformWindows: const PlatformDetails(PlatformSupport.inline),
+        },
+      ));
+
+      expect(pluginHasNativeCodeForPlatform(kPlatformLinux, plugin), isTrue);
+      expect(pluginHasNativeCodeForPlatform(kPlatformMacos, plugin), isTrue);
+      expect(pluginHasNativeCodeForPlatform(kPlatformWindows, plugin), isTrue);
+    });
+
+    test('returns true for a native+Dart plugin', () async {
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
+        'plugin',
+        packagesDir,
+        platformSupport: <String, PlatformDetails>{
+          kPlatformLinux: const PlatformDetails(PlatformSupport.inline,
+              hasNativeCode: true, hasDartCode: true),
+          kPlatformMacos: const PlatformDetails(PlatformSupport.inline,
+              hasNativeCode: true, hasDartCode: true),
+          kPlatformWindows: const PlatformDetails(PlatformSupport.inline,
+              hasNativeCode: true, hasDartCode: true),
+        },
+      ));
+
+      expect(pluginHasNativeCodeForPlatform(kPlatformLinux, plugin), isTrue);
+      expect(pluginHasNativeCodeForPlatform(kPlatformMacos, plugin), isTrue);
+      expect(pluginHasNativeCodeForPlatform(kPlatformWindows, plugin), isTrue);
+    });
+
+    test('returns false for a Dart-only plugin', () async {
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
+        'plugin',
+        packagesDir,
+        platformSupport: <String, PlatformDetails>{
+          kPlatformLinux: const PlatformDetails(PlatformSupport.inline,
+              hasNativeCode: false, hasDartCode: true),
+          kPlatformMacos: const PlatformDetails(PlatformSupport.inline,
+              hasNativeCode: false, hasDartCode: true),
+          kPlatformWindows: const PlatformDetails(PlatformSupport.inline,
+              hasNativeCode: false, hasDartCode: true),
+        },
+      ));
+
+      expect(pluginHasNativeCodeForPlatform(kPlatformLinux, plugin), isFalse);
+      expect(pluginHasNativeCodeForPlatform(kPlatformMacos, plugin), isFalse);
+      expect(pluginHasNativeCodeForPlatform(kPlatformWindows, plugin), isFalse);
     });
   });
 }
