@@ -75,11 +75,9 @@ void main() {
       );
 
       processRunner.mockProcessesForExecutable['pod'] = <io.Process>[
-        MockProcess.succeeding(),
-        MockProcess.succeeding(),
+        MockProcess(stdout: 'Foo', stderr: 'Bar'),
+        MockProcess(),
       ];
-      processRunner.resultStdout = 'Foo';
-      processRunner.resultStderr = 'Bar';
 
       final List<String> output =
           await runCapturingPrint(runner, <String>['podspecs']);
@@ -173,7 +171,7 @@ void main() {
 
       // Simulate failure from `which pod`.
       processRunner.mockProcessesForExecutable['which'] = <io.Process>[
-        MockProcess.failing(),
+        MockProcess(exitCode: 1),
       ];
 
       Error? commandError;
@@ -199,7 +197,7 @@ void main() {
 
       // Simulate failure from `pod`.
       processRunner.mockProcessesForExecutable['pod'] = <io.Process>[
-        MockProcess.failing(),
+        MockProcess(exitCode: 1),
       ];
 
       Error? commandError;
@@ -227,8 +225,8 @@ void main() {
 
       // Simulate failure from the second call to `pod`.
       processRunner.mockProcessesForExecutable['pod'] = <io.Process>[
-        MockProcess.succeeding(),
-        MockProcess.failing(),
+        MockProcess(),
+        MockProcess(exitCode: 1),
       ];
 
       Error? commandError;
