@@ -9,6 +9,7 @@ import 'package:args/command_runner.dart';
 import 'package:file/file.dart';
 import 'package:file/memory.dart';
 import 'package:flutter_plugin_tools/src/common/core.dart';
+import 'package:flutter_plugin_tools/src/common/file_utils.dart';
 import 'package:flutter_plugin_tools/src/common/plugin_utils.dart';
 import 'package:flutter_plugin_tools/src/native_test_command.dart';
 import 'package:test/test.dart';
@@ -57,7 +58,7 @@ final Map<String, dynamic> _kDeviceListMap = <String, dynamic>{
 void main() {
   const String _kDestination = '--ios-destination';
 
-  group('test native_test_command', () {
+  group('test native_test_command on Posix', () {
     late FileSystem fileSystem;
     late MockPlatform mockPlatform;
     late Directory packagesDir;
@@ -115,18 +116,16 @@ void main() {
 
     test('reports skips with no tests', () async {
       final Directory pluginDirectory1 = createFakePlugin('plugin', packagesDir,
-          platformSupport: <String, PlatformSupport>{
-            kPlatformMacos: PlatformSupport.inline,
+          platformSupport: <String, PlatformDetails>{
+            kPlatformMacos: const PlatformDetails(PlatformSupport.inline),
           });
 
       final Directory pluginExampleDirectory =
           pluginDirectory1.childDirectory('example');
 
-      // Exit code 66 from testing indicates no tests.
-      final MockProcess noTestsProcessResult = MockProcess();
-      noTestsProcessResult.exitCodeCompleter.complete(66);
       processRunner.mockProcessesForExecutable['xcrun'] = <io.Process>[
-        noTestsProcessResult,
+        // Exit code 66 from testing indicates no tests.
+        MockProcess(exitCode: 66),
       ];
       final List<String> output =
           await runCapturingPrint(runner, <String>['native-test', '--macos']);
@@ -156,8 +155,8 @@ void main() {
     group('iOS', () {
       test('skip if iOS is not supported', () async {
         createFakePlugin('plugin', packagesDir,
-            platformSupport: <String, PlatformSupport>{
-              kPlatformMacos: PlatformSupport.inline,
+            platformSupport: <String, PlatformDetails>{
+              kPlatformMacos: const PlatformDetails(PlatformSupport.inline),
             });
 
         final List<String> output = await runCapturingPrint(runner,
@@ -166,15 +165,15 @@ void main() {
             output,
             containsAllInOrder(<Matcher>[
               contains('No implementation for iOS.'),
-              contains('SKIPPING: Not implemented for target platform(s).'),
+              contains('SKIPPING: Nothing to test for target platform(s).'),
             ]));
         expect(processRunner.recordedCalls, orderedEquals(<ProcessCall>[]));
       });
 
       test('skip if iOS is implemented in a federated package', () async {
         createFakePlugin('plugin', packagesDir,
-            platformSupport: <String, PlatformSupport>{
-              kPlatformIos: PlatformSupport.federated
+            platformSupport: <String, PlatformDetails>{
+              kPlatformIos: const PlatformDetails(PlatformSupport.federated)
             });
 
         final List<String> output = await runCapturingPrint(runner,
@@ -183,15 +182,15 @@ void main() {
             output,
             containsAllInOrder(<Matcher>[
               contains('No implementation for iOS.'),
-              contains('SKIPPING: Not implemented for target platform(s).'),
+              contains('SKIPPING: Nothing to test for target platform(s).'),
             ]));
         expect(processRunner.recordedCalls, orderedEquals(<ProcessCall>[]));
       });
 
       test('running with correct destination', () async {
         final Directory pluginDirectory = createFakePlugin(
-            'plugin', packagesDir, platformSupport: <String, PlatformSupport>{
-          kPlatformIos: PlatformSupport.inline
+            'plugin', packagesDir, platformSupport: <String, PlatformDetails>{
+          kPlatformIos: const PlatformDetails(PlatformSupport.inline)
         });
 
         final Directory pluginExampleDirectory =
@@ -236,15 +235,16 @@ void main() {
       test('Not specifying --ios-destination assigns an available simulator',
           () async {
         final Directory pluginDirectory = createFakePlugin(
-            'plugin', packagesDir, platformSupport: <String, PlatformSupport>{
-          kPlatformIos: PlatformSupport.inline
+            'plugin', packagesDir, platformSupport: <String, PlatformDetails>{
+          kPlatformIos: const PlatformDetails(PlatformSupport.inline)
         });
-
         final Directory pluginExampleDirectory =
             pluginDirectory.childDirectory('example');
 
-        processRunner.processToReturn = MockProcess.succeeding();
-        processRunner.resultStdout = jsonEncode(_kDeviceListMap);
+        processRunner.mockProcessesForExecutable['xcrun'] = <io.Process>[
+          MockProcess(stdout: jsonEncode(_kDeviceListMap)), // simctl
+        ];
+
         await runCapturingPrint(runner, <String>['native-test', '--ios']);
 
         expect(
@@ -292,15 +292,15 @@ void main() {
             output,
             containsAllInOrder(<Matcher>[
               contains('No implementation for macOS.'),
-              contains('SKIPPING: Not implemented for target platform(s).'),
+              contains('SKIPPING: Nothing to test for target platform(s).'),
             ]));
         expect(processRunner.recordedCalls, orderedEquals(<ProcessCall>[]));
       });
 
       test('skip if macOS is implemented in a federated package', () async {
         createFakePlugin('plugin', packagesDir,
-            platformSupport: <String, PlatformSupport>{
-              kPlatformMacos: PlatformSupport.federated,
+            platformSupport: <String, PlatformDetails>{
+              kPlatformMacos: const PlatformDetails(PlatformSupport.federated),
             });
 
         final List<String> output =
@@ -310,7 +310,7 @@ void main() {
             output,
             containsAllInOrder(<Matcher>[
               contains('No implementation for macOS.'),
-              contains('SKIPPING: Not implemented for target platform(s).'),
+              contains('SKIPPING: Nothing to test for target platform(s).'),
             ]));
         expect(processRunner.recordedCalls, orderedEquals(<ProcessCall>[]));
       });
@@ -318,8 +318,8 @@ void main() {
       test('runs for macOS plugin', () async {
         final Directory pluginDirectory1 = createFakePlugin(
             'plugin', packagesDir,
-            platformSupport: <String, PlatformSupport>{
-              kPlatformMacos: PlatformSupport.inline,
+            platformSupport: <String, PlatformDetails>{
+              kPlatformMacos: const PlatformDetails(PlatformSupport.inline),
             });
 
         final Directory pluginExampleDirectory =
@@ -361,8 +361,8 @@ void main() {
         final Directory plugin = createFakePlugin(
           'plugin',
           packagesDir,
-          platformSupport: <String, PlatformSupport>{
-            kPlatformAndroid: PlatformSupport.inline
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.inline)
           },
           extraFiles: <String>[
             'example/android/gradlew',
@@ -391,8 +391,8 @@ void main() {
         final Directory plugin = createFakePlugin(
           'plugin',
           packagesDir,
-          platformSupport: <String, PlatformSupport>{
-            kPlatformAndroid: PlatformSupport.inline
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.inline)
           },
           extraFiles: <String>[
             'example/android/gradlew',
@@ -421,8 +421,8 @@ void main() {
         final Directory plugin = createFakePlugin(
           'plugin',
           packagesDir,
-          platformSupport: <String, PlatformSupport>{
-            kPlatformAndroid: PlatformSupport.inline
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.inline)
           },
           extraFiles: <String>[
             'example/android/gradlew',
@@ -456,8 +456,8 @@ void main() {
         createFakePlugin(
           'plugin',
           packagesDir,
-          platformSupport: <String, PlatformSupport>{
-            kPlatformAndroid: PlatformSupport.inline
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.inline)
           },
           extraFiles: <String>[
             'example/android/gradlew',
@@ -481,8 +481,8 @@ void main() {
         final Directory plugin = createFakePlugin(
           'plugin',
           packagesDir,
-          platformSupport: <String, PlatformSupport>{
-            kPlatformAndroid: PlatformSupport.inline
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.inline)
           },
           extraFiles: <String>[
             'android/src/test/example_test.java',
@@ -520,8 +520,8 @@ void main() {
         final Directory plugin = createFakePlugin(
           'plugin',
           packagesDir,
-          platformSupport: <String, PlatformSupport>{
-            kPlatformAndroid: PlatformSupport.inline
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.inline)
           },
           extraFiles: <String>[
             'android/src/test/example_test.java',
@@ -555,8 +555,8 @@ void main() {
         final Directory plugin = createFakePlugin(
           'plugin',
           packagesDir,
-          platformSupport: <String, PlatformSupport>{
-            kPlatformAndroid: PlatformSupport.inline
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.inline)
           },
           extraFiles: <String>[
             'android/src/test/example_test.java',
@@ -587,8 +587,8 @@ void main() {
         createFakePlugin(
           'plugin',
           packagesDir,
-          platformSupport: <String, PlatformSupport>{
-            kPlatformAndroid: PlatformSupport.inline
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.inline)
           },
           extraFiles: <String>[
             'example/android/app/src/test/example_test.java',
@@ -619,8 +619,8 @@ void main() {
         createFakePlugin(
           'plugin1',
           packagesDir,
-          platformSupport: <String, PlatformSupport>{
-            kPlatformAndroid: PlatformSupport.inline
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.inline)
           },
           extraFiles: <String>[
             'example/android/gradlew',
@@ -631,8 +631,8 @@ void main() {
         createFakePlugin(
           'plugin2',
           packagesDir,
-          platformSupport: <String, PlatformSupport>{
-            kPlatformAndroid: PlatformSupport.inline
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.inline)
           },
           extraFiles: <String>[
             'android/src/test/example_test.java',
@@ -658,8 +658,8 @@ void main() {
         final Directory pluginDir = createFakePlugin(
           'plugin',
           packagesDir,
-          platformSupport: <String, PlatformSupport>{
-            kPlatformAndroid: PlatformSupport.inline
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.inline)
           },
           extraFiles: <String>[
             'example/android/gradlew',
@@ -673,7 +673,7 @@ void main() {
             .childFile('gradlew')
             .path;
         processRunner.mockProcessesForExecutable[gradlewPath] = <io.Process>[
-          MockProcess.failing()
+          MockProcess(exitCode: 1)
         ];
 
         Error? commandError;
@@ -708,7 +708,7 @@ void main() {
           output,
           containsAllInOrder(<Matcher>[
             contains('No implementation for Android.'),
-            contains('SKIPPING: Not implemented for target platform(s).'),
+            contains('SKIPPING: Nothing to test for target platform(s).'),
           ]),
         );
       });
@@ -717,8 +717,8 @@ void main() {
         createFakePlugin(
           'plugin',
           packagesDir,
-          platformSupport: <String, PlatformSupport>{
-            kPlatformAndroid: PlatformSupport.inline
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.inline)
           },
         );
 
@@ -736,16 +736,157 @@ void main() {
       });
     });
 
+    group('Linux', () {
+      test('runs unit tests', () async {
+        const String testBinaryRelativePath =
+            'build/linux/foo/release/bar/plugin_test';
+        final Directory pluginDirectory =
+            createFakePlugin('plugin', packagesDir, extraFiles: <String>[
+          'example/$testBinaryRelativePath'
+        ], platformSupport: <String, PlatformDetails>{
+          kPlatformLinux: const PlatformDetails(PlatformSupport.inline),
+        });
+
+        final File testBinary = childFileWithSubcomponents(pluginDirectory,
+            <String>['example', ...testBinaryRelativePath.split('/')]);
+
+        final List<String> output = await runCapturingPrint(runner, <String>[
+          'native-test',
+          '--linux',
+          '--no-integration',
+        ]);
+
+        expect(
+          output,
+          containsAllInOrder(<Matcher>[
+            contains('Running plugin_test...'),
+            contains('No issues found!'),
+          ]),
+        );
+
+        expect(
+            processRunner.recordedCalls,
+            orderedEquals(<ProcessCall>[
+              ProcessCall(testBinary.path, const <String>[], null),
+            ]));
+      });
+
+      test('only runs release unit tests', () async {
+        const String debugTestBinaryRelativePath =
+            'build/linux/foo/debug/bar/plugin_test';
+        const String releaseTestBinaryRelativePath =
+            'build/linux/foo/release/bar/plugin_test';
+        final Directory pluginDirectory =
+            createFakePlugin('plugin', packagesDir, extraFiles: <String>[
+          'example/$debugTestBinaryRelativePath',
+          'example/$releaseTestBinaryRelativePath'
+        ], platformSupport: <String, PlatformDetails>{
+          kPlatformLinux: const PlatformDetails(PlatformSupport.inline),
+        });
+
+        final File releaseTestBinary = childFileWithSubcomponents(
+            pluginDirectory,
+            <String>['example', ...releaseTestBinaryRelativePath.split('/')]);
+
+        final List<String> output = await runCapturingPrint(runner, <String>[
+          'native-test',
+          '--linux',
+          '--no-integration',
+        ]);
+
+        expect(
+          output,
+          containsAllInOrder(<Matcher>[
+            contains('Running plugin_test...'),
+            contains('No issues found!'),
+          ]),
+        );
+
+        // Only the release version should be run.
+        expect(
+            processRunner.recordedCalls,
+            orderedEquals(<ProcessCall>[
+              ProcessCall(releaseTestBinary.path, const <String>[], null),
+            ]));
+      });
+
+      test('fails if there are no unit tests', () async {
+        createFakePlugin('plugin', packagesDir,
+            platformSupport: <String, PlatformDetails>{
+              kPlatformLinux: const PlatformDetails(PlatformSupport.inline),
+            });
+
+        Error? commandError;
+        final List<String> output = await runCapturingPrint(runner, <String>[
+          'native-test',
+          '--linux',
+          '--no-integration',
+        ], errorHandler: (Error e) {
+          commandError = e;
+        });
+
+        expect(commandError, isA<ToolExit>());
+        expect(
+          output,
+          containsAllInOrder(<Matcher>[
+            contains('No test binaries found.'),
+          ]),
+        );
+
+        expect(processRunner.recordedCalls, orderedEquals(<ProcessCall>[]));
+      });
+
+      test('fails if a unit test fails', () async {
+        const String testBinaryRelativePath =
+            'build/linux/foo/release/bar/plugin_test';
+        final Directory pluginDirectory =
+            createFakePlugin('plugin', packagesDir, extraFiles: <String>[
+          'example/$testBinaryRelativePath'
+        ], platformSupport: <String, PlatformDetails>{
+          kPlatformLinux: const PlatformDetails(PlatformSupport.inline),
+        });
+
+        final File testBinary = childFileWithSubcomponents(pluginDirectory,
+            <String>['example', ...testBinaryRelativePath.split('/')]);
+
+        processRunner.mockProcessesForExecutable[testBinary.path] =
+            <io.Process>[MockProcess(exitCode: 1)];
+
+        Error? commandError;
+        final List<String> output = await runCapturingPrint(runner, <String>[
+          'native-test',
+          '--linux',
+          '--no-integration',
+        ], errorHandler: (Error e) {
+          commandError = e;
+        });
+
+        expect(commandError, isA<ToolExit>());
+        expect(
+          output,
+          containsAllInOrder(<Matcher>[
+            contains('Running plugin_test...'),
+          ]),
+        );
+
+        expect(
+            processRunner.recordedCalls,
+            orderedEquals(<ProcessCall>[
+              ProcessCall(testBinary.path, const <String>[], null),
+            ]));
+      });
+    });
+
     // Tests behaviors of implementation that is shared between iOS and macOS.
     group('iOS/macOS', () {
       test('fails if xcrun fails', () async {
         createFakePlugin('plugin', packagesDir,
-            platformSupport: <String, PlatformSupport>{
-              kPlatformMacos: PlatformSupport.inline,
+            platformSupport: <String, PlatformDetails>{
+              kPlatformMacos: const PlatformDetails(PlatformSupport.inline),
             });
 
         processRunner.mockProcessesForExecutable['xcrun'] = <io.Process>[
-          MockProcess.failing()
+          MockProcess(exitCode: 1)
         ];
 
         Error? commandError;
@@ -768,16 +909,21 @@ void main() {
       test('honors unit-only', () async {
         final Directory pluginDirectory1 = createFakePlugin(
             'plugin', packagesDir,
-            platformSupport: <String, PlatformSupport>{
-              kPlatformMacos: PlatformSupport.inline,
+            platformSupport: <String, PlatformDetails>{
+              kPlatformMacos: const PlatformDetails(PlatformSupport.inline),
             });
 
         final Directory pluginExampleDirectory =
             pluginDirectory1.childDirectory('example');
 
-        processRunner.processToReturn = MockProcess.succeeding();
-        processRunner.resultStdout =
-            '{"project":{"targets":["RunnerTests", "RunnerUITests"]}}';
+        const Map<String, dynamic> projects = <String, dynamic>{
+          'project': <String, dynamic>{
+            'targets': <String>['RunnerTests', 'RunnerUITests']
+          }
+        };
+        processRunner.mockProcessesForExecutable['xcrun'] = <io.Process>[
+          MockProcess(stdout: jsonEncode(projects)), // xcodebuild -list
+        ];
 
         final List<String> output = await runCapturingPrint(runner, <String>[
           'native-test',
@@ -828,16 +974,21 @@ void main() {
       test('honors integration-only', () async {
         final Directory pluginDirectory1 = createFakePlugin(
             'plugin', packagesDir,
-            platformSupport: <String, PlatformSupport>{
-              kPlatformMacos: PlatformSupport.inline,
+            platformSupport: <String, PlatformDetails>{
+              kPlatformMacos: const PlatformDetails(PlatformSupport.inline),
             });
 
         final Directory pluginExampleDirectory =
             pluginDirectory1.childDirectory('example');
 
-        processRunner.processToReturn = MockProcess.succeeding();
-        processRunner.resultStdout =
-            '{"project":{"targets":["RunnerTests", "RunnerUITests"]}}';
+        const Map<String, dynamic> projects = <String, dynamic>{
+          'project': <String, dynamic>{
+            'targets': <String>['RunnerTests', 'RunnerUITests']
+          }
+        };
+        processRunner.mockProcessesForExecutable['xcrun'] = <io.Process>[
+          MockProcess(stdout: jsonEncode(projects)), // xcodebuild -list
+        ];
 
         final List<String> output = await runCapturingPrint(runner, <String>[
           'native-test',
@@ -888,16 +1039,23 @@ void main() {
       test('skips when the requested target is not present', () async {
         final Directory pluginDirectory1 = createFakePlugin(
             'plugin', packagesDir,
-            platformSupport: <String, PlatformSupport>{
-              kPlatformMacos: PlatformSupport.inline,
+            platformSupport: <String, PlatformDetails>{
+              kPlatformMacos: const PlatformDetails(PlatformSupport.inline),
             });
 
         final Directory pluginExampleDirectory =
             pluginDirectory1.childDirectory('example');
 
-        processRunner.processToReturn = MockProcess.succeeding();
         // Simulate a project with unit tests but no integration tests...
-        processRunner.resultStdout = '{"project":{"targets":["RunnerTests"]}}';
+        const Map<String, dynamic> projects = <String, dynamic>{
+          'project': <String, dynamic>{
+            'targets': <String>['RunnerTests']
+          }
+        };
+        processRunner.mockProcessesForExecutable['xcrun'] = <io.Process>[
+          MockProcess(stdout: jsonEncode(projects)), // xcodebuild -list
+        ];
+
         // ... then try to run only integration tests.
         final List<String> output = await runCapturingPrint(runner, <String>[
           'native-test',
@@ -934,14 +1092,16 @@ void main() {
       test('fails if unable to check for requested target', () async {
         final Directory pluginDirectory1 = createFakePlugin(
             'plugin', packagesDir,
-            platformSupport: <String, PlatformSupport>{
-              kPlatformMacos: PlatformSupport.inline,
+            platformSupport: <String, PlatformDetails>{
+              kPlatformMacos: const PlatformDetails(PlatformSupport.inline),
             });
 
         final Directory pluginExampleDirectory =
             pluginDirectory1.childDirectory('example');
 
-        processRunner.processToReturn = MockProcess.failing();
+        processRunner.mockProcessesForExecutable['xcrun'] = <io.Process>[
+          MockProcess(exitCode: 1), // xcodebuild -list
+        ];
 
         Error? commandError;
         final List<String> output = await runCapturingPrint(runner, <String>[
@@ -989,10 +1149,10 @@ void main() {
             'example/android/gradlew',
             'android/src/test/example_test.java',
           ],
-          platformSupport: <String, PlatformSupport>{
-            kPlatformAndroid: PlatformSupport.inline,
-            kPlatformIos: PlatformSupport.inline,
-            kPlatformMacos: PlatformSupport.inline,
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.inline),
+            kPlatformIos: const PlatformDetails(PlatformSupport.inline),
+            kPlatformMacos: const PlatformDetails(PlatformSupport.inline),
           },
         );
 
@@ -1059,8 +1219,8 @@ void main() {
       test('runs only macOS for a macOS plugin', () async {
         final Directory pluginDirectory1 = createFakePlugin(
             'plugin', packagesDir,
-            platformSupport: <String, PlatformSupport>{
-              kPlatformMacos: PlatformSupport.inline,
+            platformSupport: <String, PlatformDetails>{
+              kPlatformMacos: const PlatformDetails(PlatformSupport.inline),
             });
 
         final Directory pluginExampleDirectory =
@@ -1103,8 +1263,8 @@ void main() {
 
       test('runs only iOS for a iOS plugin', () async {
         final Directory pluginDirectory = createFakePlugin(
-            'plugin', packagesDir, platformSupport: <String, PlatformSupport>{
-          kPlatformIos: PlatformSupport.inline
+            'plugin', packagesDir, platformSupport: <String, PlatformDetails>{
+          kPlatformIos: const PlatformDetails(PlatformSupport.inline)
         });
 
         final Directory pluginExampleDirectory =
@@ -1155,6 +1315,7 @@ void main() {
           '--android',
           '--ios',
           '--macos',
+          '--windows',
           _kDestination,
           'foo_destination',
         ]);
@@ -1165,7 +1326,38 @@ void main() {
               contains('No implementation for Android.'),
               contains('No implementation for iOS.'),
               contains('No implementation for macOS.'),
-              contains('SKIPPING: Not implemented for target platform(s).'),
+              contains('SKIPPING: Nothing to test for target platform(s).'),
+            ]));
+
+        expect(processRunner.recordedCalls, orderedEquals(<ProcessCall>[]));
+      });
+
+      test('skips Dart-only plugins', () async {
+        createFakePlugin(
+          'plugin',
+          packagesDir,
+          platformSupport: <String, PlatformDetails>{
+            kPlatformMacos: const PlatformDetails(PlatformSupport.inline,
+                hasDartCode: true, hasNativeCode: false),
+            kPlatformWindows: const PlatformDetails(PlatformSupport.inline,
+                hasDartCode: true, hasNativeCode: false),
+          },
+        );
+
+        final List<String> output = await runCapturingPrint(runner, <String>[
+          'native-test',
+          '--macos',
+          '--windows',
+          _kDestination,
+          'foo_destination',
+        ]);
+
+        expect(
+            output,
+            containsAllInOrder(<Matcher>[
+              contains('No native code for macOS.'),
+              contains('No native code for Windows.'),
+              contains('SKIPPING: Nothing to test for target platform(s).'),
             ]));
 
         expect(processRunner.recordedCalls, orderedEquals(<ProcessCall>[]));
@@ -1175,9 +1367,9 @@ void main() {
         final Directory pluginDir = createFakePlugin(
           'plugin',
           packagesDir,
-          platformSupport: <String, PlatformSupport>{
-            kPlatformAndroid: PlatformSupport.inline,
-            kPlatformIos: PlatformSupport.inline,
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.inline),
+            kPlatformIos: const PlatformDetails(PlatformSupport.inline),
           },
           extraFiles: <String>[
             'example/android/gradlew',
@@ -1192,7 +1384,7 @@ void main() {
             .childFile('gradlew')
             .path;
         processRunner.mockProcessesForExecutable[gradlewPath] = <io.Process>[
-          MockProcess.failing()
+          MockProcess(exitCode: 1)
         ];
 
         Error? commandError;
@@ -1226,9 +1418,9 @@ void main() {
         final Directory pluginDir = createFakePlugin(
           'plugin',
           packagesDir,
-          platformSupport: <String, PlatformSupport>{
-            kPlatformAndroid: PlatformSupport.inline,
-            kPlatformIos: PlatformSupport.inline,
+          platformSupport: <String, PlatformDetails>{
+            kPlatformAndroid: const PlatformDetails(PlatformSupport.inline),
+            kPlatformIos: const PlatformDetails(PlatformSupport.inline),
           },
           extraFiles: <String>[
             'example/android/gradlew',
@@ -1243,11 +1435,11 @@ void main() {
             .childFile('gradlew')
             .path;
         processRunner.mockProcessesForExecutable[gradlewPath] = <io.Process>[
-          MockProcess.failing()
+          MockProcess(exitCode: 1)
         ];
         // Simulate failing Android.
         processRunner.mockProcessesForExecutable['xcrun'] = <io.Process>[
-          MockProcess.failing()
+          MockProcess(exitCode: 1)
         ];
 
         Error? commandError;
@@ -1274,6 +1466,168 @@ void main() {
                 '    iOS')
           ]),
         );
+      });
+    });
+  });
+
+  group('test native_test_command on Windows', () {
+    late FileSystem fileSystem;
+    late MockPlatform mockPlatform;
+    late Directory packagesDir;
+    late CommandRunner<void> runner;
+    late RecordingProcessRunner processRunner;
+
+    setUp(() {
+      fileSystem = MemoryFileSystem(style: FileSystemStyle.windows);
+      mockPlatform = MockPlatform(isWindows: true);
+      packagesDir = createPackagesDirectory(fileSystem: fileSystem);
+      processRunner = RecordingProcessRunner();
+      final NativeTestCommand command = NativeTestCommand(packagesDir,
+          processRunner: processRunner, platform: mockPlatform);
+
+      runner = CommandRunner<void>(
+          'native_test_command', 'Test for native_test_command');
+      runner.addCommand(command);
+    });
+
+    group('Windows', () {
+      test('runs unit tests', () async {
+        const String testBinaryRelativePath =
+            'build/windows/foo/Release/bar/plugin_test.exe';
+        final Directory pluginDirectory =
+            createFakePlugin('plugin', packagesDir, extraFiles: <String>[
+          'example/$testBinaryRelativePath'
+        ], platformSupport: <String, PlatformDetails>{
+          kPlatformWindows: const PlatformDetails(PlatformSupport.inline),
+        });
+
+        final File testBinary = childFileWithSubcomponents(pluginDirectory,
+            <String>['example', ...testBinaryRelativePath.split('/')]);
+
+        final List<String> output = await runCapturingPrint(runner, <String>[
+          'native-test',
+          '--windows',
+          '--no-integration',
+        ]);
+
+        expect(
+          output,
+          containsAllInOrder(<Matcher>[
+            contains('Running plugin_test.exe...'),
+            contains('No issues found!'),
+          ]),
+        );
+
+        expect(
+            processRunner.recordedCalls,
+            orderedEquals(<ProcessCall>[
+              ProcessCall(testBinary.path, const <String>[], null),
+            ]));
+      });
+
+      test('only runs release unit tests', () async {
+        const String debugTestBinaryRelativePath =
+            'build/windows/foo/Debug/bar/plugin_test.exe';
+        const String releaseTestBinaryRelativePath =
+            'build/windows/foo/Release/bar/plugin_test.exe';
+        final Directory pluginDirectory =
+            createFakePlugin('plugin', packagesDir, extraFiles: <String>[
+          'example/$debugTestBinaryRelativePath',
+          'example/$releaseTestBinaryRelativePath'
+        ], platformSupport: <String, PlatformDetails>{
+          kPlatformWindows: const PlatformDetails(PlatformSupport.inline),
+        });
+
+        final File releaseTestBinary = childFileWithSubcomponents(
+            pluginDirectory,
+            <String>['example', ...releaseTestBinaryRelativePath.split('/')]);
+
+        final List<String> output = await runCapturingPrint(runner, <String>[
+          'native-test',
+          '--windows',
+          '--no-integration',
+        ]);
+
+        expect(
+          output,
+          containsAllInOrder(<Matcher>[
+            contains('Running plugin_test.exe...'),
+            contains('No issues found!'),
+          ]),
+        );
+
+        // Only the release version should be run.
+        expect(
+            processRunner.recordedCalls,
+            orderedEquals(<ProcessCall>[
+              ProcessCall(releaseTestBinary.path, const <String>[], null),
+            ]));
+      });
+
+      test('fails if there are no unit tests', () async {
+        createFakePlugin('plugin', packagesDir,
+            platformSupport: <String, PlatformDetails>{
+              kPlatformWindows: const PlatformDetails(PlatformSupport.inline),
+            });
+
+        Error? commandError;
+        final List<String> output = await runCapturingPrint(runner, <String>[
+          'native-test',
+          '--windows',
+          '--no-integration',
+        ], errorHandler: (Error e) {
+          commandError = e;
+        });
+
+        expect(commandError, isA<ToolExit>());
+        expect(
+          output,
+          containsAllInOrder(<Matcher>[
+            contains('No test binaries found.'),
+          ]),
+        );
+
+        expect(processRunner.recordedCalls, orderedEquals(<ProcessCall>[]));
+      });
+
+      test('fails if a unit test fails', () async {
+        const String testBinaryRelativePath =
+            'build/windows/foo/Release/bar/plugin_test.exe';
+        final Directory pluginDirectory =
+            createFakePlugin('plugin', packagesDir, extraFiles: <String>[
+          'example/$testBinaryRelativePath'
+        ], platformSupport: <String, PlatformDetails>{
+          kPlatformWindows: const PlatformDetails(PlatformSupport.inline),
+        });
+
+        final File testBinary = childFileWithSubcomponents(pluginDirectory,
+            <String>['example', ...testBinaryRelativePath.split('/')]);
+
+        processRunner.mockProcessesForExecutable[testBinary.path] =
+            <io.Process>[MockProcess(exitCode: 1)];
+
+        Error? commandError;
+        final List<String> output = await runCapturingPrint(runner, <String>[
+          'native-test',
+          '--windows',
+          '--no-integration',
+        ], errorHandler: (Error e) {
+          commandError = e;
+        });
+
+        expect(commandError, isA<ToolExit>());
+        expect(
+          output,
+          containsAllInOrder(<Matcher>[
+            contains('Running plugin_test.exe...'),
+          ]),
+        );
+
+        expect(
+            processRunner.recordedCalls,
+            orderedEquals(<ProcessCall>[
+              ProcessCall(testBinary.path, const <String>[], null),
+            ]));
       });
     });
   });
