@@ -67,8 +67,8 @@ void main() {
       processRunner
               .mockProcessesForExecutable[getFlutterCommand(mockPlatform)] =
           <io.Process>[
-        MockProcess.failing(), // plugin 1 test
-        MockProcess.succeeding(), // plugin 2 test
+        MockProcess(exitCode: 1), // plugin 1 test
+        MockProcess(), // plugin 2 test
       ];
 
       Error? commandError;
@@ -132,7 +132,7 @@ void main() {
           extraFiles: <String>['test/empty_test.dart']);
 
       processRunner.mockProcessesForExecutable['dart'] = <io.Process>[
-        MockProcess.failing(), // dart pub get
+        MockProcess(exitCode: 1), // dart pub get
       ];
 
       Error? commandError;
@@ -156,8 +156,8 @@ void main() {
           extraFiles: <String>['test/empty_test.dart']);
 
       processRunner.mockProcessesForExecutable['dart'] = <io.Process>[
-        MockProcess.succeeding(), // dart pub get
-        MockProcess.failing(), // dart pub run test
+        MockProcess(), // dart pub get
+        MockProcess(exitCode: 1), // dart pub run test
       ];
 
       Error? commandError;
@@ -180,8 +180,8 @@ void main() {
         'plugin',
         packagesDir,
         extraFiles: <String>['test/empty_test.dart'],
-        platformSupport: <String, PlatformSupport>{
-          kPlatformWeb: PlatformSupport.inline,
+        platformSupport: <String, PlatformDetails>{
+          kPlatformWeb: const PlatformDetails(PlatformSupport.inline),
         },
       );
 
