@@ -10,12 +10,15 @@ import 'package:args/command_runner.dart';
 import 'package:file/file.dart';
 import 'package:file/memory.dart';
 import 'package:flutter_plugin_tools/src/common/core.dart';
+import 'package:flutter_plugin_tools/src/common/file_utils.dart';
 import 'package:flutter_plugin_tools/src/common/plugin_utils.dart';
 import 'package:flutter_plugin_tools/src/common/process_runner.dart';
 import 'package:meta/meta.dart';
 import 'package:path/path.dart' as p;
 import 'package:platform/platform.dart';
 import 'package:quiver/collection.dart';
+
+import 'mocks.dart';
 
 /// Returns the exe name that command will use when running Flutter on
 /// [platform].
@@ -39,6 +42,33 @@ Directory createPackagesDirectory(
   return packagesDir;
 }
 
+/// Details for platform support in a plugin.
+@immutable
+class PlatformDetails {
+  const PlatformDetails(
+    this.type, {
+    this.variants = const <String>[],
+    this.hasNativeCode = true,
+    this.hasDartCode = false,
+  });
+
+  /// The type of support for the platform.
+  final PlatformSupport type;
+
+  /// Any 'supportVariants' to list in the pubspec.
+  final List<String> variants;
+
+  /// Whether or not the plugin includes native code.
+  ///
+  /// Ignored for web, which does not have native code.
+  final bool hasNativeCode;
+
+  /// Whether or not the plugin includes Dart code.
+  ///
+  /// Ignored for web, which always has native code.
+  final bool hasDartCode;
+}
+
 /// Creates a plugin package with the given [name] in [packagesDirectory].
 ///
 /// [platformSupport] is a map of platform string to the support details for
@@ -46,13 +76,14 @@ Directory createPackagesDirectory(
 ///
 /// [extraFiles] is an optional list of plugin-relative paths, using Posix
 /// separators, of extra files to create in the plugin.
+// TODO(stuartmorgan): Convert the return to a RepositoryPackage.
 Directory createFakePlugin(
   String name,
   Directory parentDirectory, {
   List<String> examples = const <String>['example'],
   List<String> extraFiles = const <String>[],
-  Map<String, PlatformSupport> platformSupport =
-      const <String, PlatformSupport>{},
+  Map<String, PlatformDetails> platformSupport =
+      const <String, PlatformDetails>{},
   String? version = '0.0.1',
 }) {
   final Directory pluginDirectory = createFakePackage(name, parentDirectory,
@@ -77,6 +108,7 @@ Directory createFakePlugin(
 ///
 /// [extraFiles] is an optional list of package-relative paths, using unix-style
 /// separators, of extra files to create in the package.
+// TODO(stuartmorgan): Convert the return to a RepositoryPackage.
 Directory createFakePackage(
   String name,
   Directory parentDirectory, {
@@ -88,7 +120,8 @@ Directory createFakePackage(
   final Directory packageDirectory = parentDirectory.childDirectory(name);
   packageDirectory.createSync(recursive: true);
 
-  createFakePubspec(packageDirectory, name: name, isFlutter: isFlutter);
+  createFakePubspec(packageDirectory,
+      name: name, isFlutter: isFlutter, version: version);
   createFakeCHANGELOG(packageDirectory, '''
 ## $version
   * Some changes.
@@ -110,15 +143,10 @@ Directory createFakePackage(
     }
   }
 
-  final FileSystem fileSystem = packageDirectory.fileSystem;
   final p.Context posixContext = p.posix;
   for (final String file in extraFiles) {
-    final List<String> newFilePath = <String>[
-      packageDirectory.path,
-      ...posixContext.split(file)
-    ];
-    final File newFile = fileSystem.file(fileSystem.path.joinAll(newFilePath));
-    newFile.createSync(recursive: true);
+    childFileWithSubcomponents(packageDirectory, posixContext.split(file))
+        .createSync(recursive: true);
   }
 
   return packageDirectory;
@@ -139,8 +167,8 @@ void createFakePubspec(
   String name = 'fake_package',
   bool isFlutter = true,
   bool isPlugin = false,
-  Map<String, PlatformSupport> platformSupport =
-      const <String, PlatformSupport>{},
+  Map<String, PlatformDetails> platformSupport =
+      const <String, PlatformDetails>{},
   String publishTo = 'http://no_pub_server.com',
   String? version,
 }) {
@@ -156,12 +184,11 @@ flutter:
   plugin:
     platforms:
 ''';
-      for (final MapEntry<String, PlatformSupport> platform
+      for (final MapEntry<String, PlatformDetails> platform
           in platformSupport.entries) {
         yaml += _pluginPlatformSection(platform.key, platform.value, name);
       }
     }
-
     yaml += '''
 dependencies:
   flutter:
@@ -182,50 +209,62 @@ publish_to: $publishTo # Hardcoded safeguard to prevent this from somehow being 
 }
 
 String _pluginPlatformSection(
-    String platform, PlatformSupport type, String packageName) {
-  if (type == PlatformSupport.federated) {
-    return '''
+    String platform, PlatformDetails support, String packageName) {
+  String entry = '';
+  // Build the main plugin entry.
+  if (support.type == PlatformSupport.federated) {
+    entry = '''
       $platform:
         default_package: ${packageName}_$platform
 ''';
+  } else {
+    final List<String> lines = <String>[
+      '      $platform:',
+    ];
+    switch (platform) {
+      case kPlatformAndroid:
+        lines.add('        package: io.flutter.plugins.fake');
+        continue nativeByDefault;
+      nativeByDefault:
+      case kPlatformIos:
+      case kPlatformLinux:
+      case kPlatformMacos:
+      case kPlatformWindows:
+        if (support.hasNativeCode) {
+          final String className =
+              platform == kPlatformIos ? 'FLTFakePlugin' : 'FakePlugin';
+          lines.add('        pluginClass: $className');
+        }
+        if (support.hasDartCode) {
+          lines.add('        dartPluginClass: FakeDartPlugin');
+        }
+        break;
+      case kPlatformWeb:
+        lines.addAll(<String>[
+          '        pluginClass: FakePlugin',
+          '        fileName: ${packageName}_web.dart',
+        ]);
+        break;
+      default:
+        assert(false, 'Unrecognized platform: $platform');
+        break;
+    }
+    entry = lines.join('\n') + '\n';
   }
-  switch (platform) {
-    case kPlatformAndroid:
-      return '''
-      android:
-        package: io.flutter.plugins.fake
-        pluginClass: FakePlugin
+
+  // Add any variants.
+  if (support.variants.isNotEmpty) {
+    entry += '''
+        supportedVariants:
 ''';
-    case kPlatformIos:
-      return '''
-      ios:
-        pluginClass: FLTFakePlugin
+    for (final String variant in support.variants) {
+      entry += '''
+          - $variant
 ''';
-    case kPlatformLinux:
-      return '''
-      linux:
-        pluginClass: FakePlugin
-''';
-    case kPlatformMacos:
-      return '''
-      macos:
-        pluginClass: FakePlugin
-''';
-    case kPlatformWeb:
-      return '''
-      web:
-        pluginClass: FakePlugin
-        fileName: ${packageName}_web.dart
-''';
-    case kPlatformWindows:
-      return '''
-      windows:
-        pluginClass: FakePlugin
-''';
-    default:
-      assert(false);
-      return '';
+    }
   }
+
+  return entry;
 }
 
 typedef _ErrorHandler = void Function(Error error);
@@ -265,15 +304,6 @@ class RecordingProcessRunner extends ProcessRunner {
   final Map<String, List<io.Process>> mockProcessesForExecutable =
       <String, List<io.Process>>{};
 
-  /// Populate for [io.ProcessResult] to use a String [stdout] instead of a [List] of [int].
-  String? resultStdout;
-
-  /// Populate for [io.ProcessResult] to use a String [stderr] instead of a [List] of [int].
-  String? resultStderr;
-
-  // Deprecated--do not add new uses. Use mockProcessesForExecutable instead.
-  io.Process? processToReturn;
-
   @override
   Future<int> runAndStream(
     String executable,
@@ -291,8 +321,7 @@ class RecordingProcessRunner extends ProcessRunner {
     return Future<int>.value(exitCode);
   }
 
-  /// Returns [io.ProcessResult] created from [mockProcessesForExecutable],
-  /// [resultStdout], and [resultStderr].
+  /// Returns [io.ProcessResult] created from [mockProcessesForExecutable].
   @override
   Future<io.ProcessResult> run(
     String executable,
@@ -306,10 +335,16 @@ class RecordingProcessRunner extends ProcessRunner {
     recordedCalls.add(ProcessCall(executable, args, workingDir?.path));
 
     final io.Process? process = _getProcessToReturn(executable);
+    final List<String>? processStdout =
+        await process?.stdout.transform(stdoutEncoding.decoder).toList();
+    final String stdout = processStdout?.join('') ?? '';
+    final List<String>? processStderr =
+        await process?.stderr.transform(stderrEncoding.decoder).toList();
+    final String stderr = processStderr?.join('') ?? '';
+
     final io.ProcessResult result = process == null
         ? io.ProcessResult(1, 0, '', '')
-        : io.ProcessResult(process.pid, await process.exitCode,
-            resultStdout ?? process.stdout, resultStderr ?? process.stderr);
+        : io.ProcessResult(process.pid, await process.exitCode, stdout, stderr);
 
     if (exitOnError && (result.exitCode != 0)) {
       throw io.ProcessException(executable, args);
@@ -322,17 +357,16 @@ class RecordingProcessRunner extends ProcessRunner {
   Future<io.Process> start(String executable, List<String> args,
       {Directory? workingDirectory}) async {
     recordedCalls.add(ProcessCall(executable, args, workingDirectory?.path));
-    return Future<io.Process>.value(_getProcessToReturn(executable));
+    return Future<io.Process>.value(
+        _getProcessToReturn(executable) ?? MockProcess());
   }
 
   io.Process? _getProcessToReturn(String executable) {
-    io.Process? process;
     final List<io.Process>? processes = mockProcessesForExecutable[executable];
     if (processes != null && processes.isNotEmpty) {
-      process = mockProcessesForExecutable[executable]!.removeAt(0);
+      return processes.removeAt(0);
     }
-    // Fall back to `processToReturn` for backwards compatibility.
-    return process ?? processToReturn;
+    return null;
   }
 }
 
